# backend/cmd/api/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_NeroQue_course-management-backend_internal_models.Profile:
    properties:
      created_at:
        allOf:
        - $ref: '#/definitions/sql.NullTime'
        description: timestamps
      experience:
        description: gamification stuff
        type: integer
      gems:
        description: special currency
        type: integer
      id:
        description: unique identifier
        type: string
      last_active_date:
        allOf:
        - $ref: '#/definitions/sql.NullTime'
        description: for streak tracking
      name:
        description: display name
        type: string
      streak:
        description: consecutive active days
        type: integer
      updated_at:
        $ref: '#/definitions/sql.NullTime'
    type: object
...
  sql.NullTime:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing courses, modules, profiles, and user progress in a
    learning management system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Course Management System API
  version: "1.0"
paths:
...
  /profiles:
    delete:
      consumes:
      - application/json
      description: Delete a user profile by ID
      parameters:
      - description: Delete request with user_id
        in: body
        name: deleteRequest
        required: true
        schema:
          properties:
            user_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Profile deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format or missing user ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a profile
      tags:
      - profiles
    get:
      consumes:
      - application/json
      description: Get a list of all user profiles in the system
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved profiles
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List all profiles
      tags:
      - profiles
    post:
      consumes:
      - application/json
      description: Create a new user profile with the provided name
      parameters:
      - description: Profile object with name field
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/github_com_NeroQue_course-management-backend_internal_models.Profile'
      produces:
      - application/json
      responses:
        "201":
          description: Profile created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format or missing required fields
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new profile
      tags:
      - profiles
    put:
      consumes:
      - application/json
      description: Update an existing profile's name
      parameters:
      - description: Update request with user_id and new_name
        in: body
        name: updateRequest
        required: true
        schema:
          properties:
            new_name:
              type: string
            user_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format or missing required fields
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update a profile
      tags:
      - profiles
  /profiles/{id}/select:
    post:
      consumes:
      - application/json
      description: Set a profile as the currently active user profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Profile selected successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid profile ID format
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Profile not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Select an active profile
      tags:
      - profiles
  /tasks:
    get:
      consumes:
      - application/json
      description: Get the current status and result of a background task
      parameters:
      - description: Task ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task status and information
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing or invalid task ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties: true
            type: object
      summary: Get task status
      tags:
      - tasks
  /tasks/cleanup:
    post:
      consumes:
      - application/json
      description: Manually trigger cleanup of old completed tasks
      parameters:
      - description: Age duration (e.g., '24h', '1h30m'). Defaults to 24h
        in: query
        name: age
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Number of tasks cleaned up
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid duration format
          schema:
            additionalProperties: true
            type: object
      summary: Cleanup old tasks
      tags:
      - tasks
...
schemes:
- http
- https
securityDefinitions:
  SessionAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"